% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellChatHelper.R
\name{CellChatHelper}
\alias{CellChatHelper}
\title{Main function of CellChat analysis}
\usage{
CellChatHelper(
  seu,
  label.field,
  name = "name",
  mode = "default",
  DB = NULL,
  out.dir = getwd(),
  cores = 10,
  fig.width = 10,
  fig.height = 8
)
}
\arguments{
\item{seu}{seurat object}

\item{label.field}{filed name in seurat meta.data contains cell labels.}

\item{name}{run name, the prefix of final rds file and pdf plot.}

\item{mode}{one of default, sensitive, population.
- default: default CellChat settings.
- sensitive: sensitive to low expressed genes.
- population: whether consider the proportion of cells in each group across all sequenced cells.}

\item{DB}{cellchat database. Default: NULL}

\item{out.dir}{output path, including pdf plots and cellchat rds file. Default: ./}

\item{cores}{threads used}

\item{fig.width}{figure width for pdf file. Default: 10}

\item{fig.height}{figure height for pdf file. Default: 8}
}
\description{
Main function of CellChat analysis
}
