% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichmentAnalysis.R
\name{enrich_batch}
\alias{enrich_batch}
\title{Perform Batch Enrichment Analysis}
\usage{
enrich_batch(all.markers, t2g)
}
\arguments{
\item{all.markers}{A list of marker sets, where each element is a vector of gene identifiers.}

\item{t2g}{A data frame with at least two columns: term and gene, for the enrichment analysis.}
}
\value{
A list of enrichment results (clusterProfiler::enricher object), with each element corresponding to an input marker set.
         NULL results are removed from the output.
}
\description{
This function performs enrichment analysis on a list of marker sets using the clusterProfiler package.
}
\examples{
# Example usage:
# result <- enrich_batch(my_markers, my_term2gene)

}
