% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelection.R
\name{FindIntVarFeatures}
\alias{FindIntVarFeatures}
\title{Find Variable Features from Integrated Dataset}
\usage{
FindIntVarFeatures(
  object,
  split.by,
  do.norm = TRUE,
  selection.method = "vst",
  nfeatures = 2000,
  min.cells = 500,
  ...
)
}
\arguments{
\item{object}{A Seurat object.}

\item{split.by}{The grouping variable used to split the input object.}

\item{do.norm}{A logical indicating whether the data should be normalized. Default is TRUE.}

\item{selection.method}{The method used to select variable features. Default is "vst".}

\item{nfeatures}{The number of variable features to select. Default is 2000.}

\item{min.cells}{The number of minimal cells for calculation. Default is 500.}

\item{...}{Additional arguments passed to FindVariableFeatures.}
}
\value{
A vector of selected feature names.
}
\description{
Find Variable Features from Integrated Dataset
}
\examples{
\dontrun{
# Find variable features
vfeatures <- FindIntVarFeatures(object = seu, split.by = "Batch")
}

}
