% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_label_pos}
\alias{get_label_pos}
\title{calculating the position of cluster labels}
\usage{
get_label_pos(data, emb = "tSNE", group.by = "ClusterID")
}
\arguments{
\item{data}{A data.frame contains embeddings and clusters.}

\item{emb}{A string indicating the prefix of embedding names. Default: "tSNE"}

\item{group.by}{A string indicating the field of clusters. Default: "ClusterID"}
}
\value{
A data.frame contains median coordinates of each cluster.
}
\description{
calculating the position of cluster labels
}
