% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{impute_nmf}
\alias{impute_nmf}
\title{Impute gene expression data using Non-negative Matrix Factorization (NMF)}
\usage{
impute_nmf(seu, min_cells = 50, k = 100, threads = 1, seed = 1024)
}
\arguments{
\item{seu}{A Seurat object containing the RNA assay with raw counts.}

\item{min_cells}{Minimum number of cells a gene must be expressed in to be included. Default is 50.}

\item{k}{Number of factors for NMF. Default is 100.}

\item{threads}{Number of threads to use for parallel computation. Default is 1.}

\item{seed}{Random seed for reproducibility. Default is 1024.}
}
\value{
A Seurat object with an additional "imputed" assay containing the imputed data.
}
\description{
This function performs data imputation on a Seurat object using NMF.
}
\details{
The function performs the following steps:
1. Filters genes based on the minimum number of cells they are expressed in.
2. Extracts the log-normalized expression matrix from the RNA assay.
3. Applies NMF using the RcppML package.
4. Reconstructs the imputed matrix from the NMF factors.
5. Adds the imputed matrix as a new assay to the Seurat object.
}
