% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{ReadSolo}
\alias{ReadSolo}
\title{Read output from STARsolo}
\usage{
ReadSolo(path, assay = "filtered")
}
\arguments{
\item{path}{Directory contains sample dirs. Sample dirs containing data files.}

\item{assay}{Load raw or filtered matrix generated by STARsolo. Default: "filtered"}
}
\value{
A single sparse matrix
}
\description{
Read output from STARsolo
}
