% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoveAmbientRNAs.R
\name{RemoveAmbientRNAs}
\alias{RemoveAmbientRNAs}
\title{A wrapper of celda::decontX}
\usage{
RemoveAmbientRNAs(seu, cluster.name = NULL, library.id = NULL)
}
\arguments{
\item{seu}{Seurat object}

\item{cluster.name}{Name of cluster field in Seurat object.}

\item{library.id}{Name of library field in Seurat object.}
}
\value{
seurat object including a corrected counts matrix and contamination rates for each cell.
}
\description{
A wrapper of celda::decontX
}
