% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_mkdir}
\alias{safe_mkdir}
\title{Make directory if it not exists.}
\usage{
safe_mkdir(path)
}
\arguments{
\item{path}{A character vector containing a single path name}
}
\description{
Make directory if it not exists.
}
