% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkDoublets.R
\name{MarkDoublets}
\alias{MarkDoublets}
\title{Mark doublets}
\usage{
MarkDoublets(seu, PCs = 1:10, split.by = NULL, num.cores = 1)
}
\arguments{
\item{seu}{Seurat object}

\item{PCs}{Vectors indicating used principle components. Default: 1:10}

\item{split.by}{Name of a metadata column to split plot by. Default: NULL}

\item{num.cores}{Threads for calculation. Default: 1}
}
\value{
Seurat object
}
\description{
Mark doublets
}
