% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoveAmbientRNAs.R
\name{RemoveAmbientRNAs}
\alias{RemoveAmbientRNAs}
\title{A wrapper of celda::decontX}
\usage{
RemoveAmbientRNAs(seu, split.by = NULL, cluster.name = NULL)
}
\arguments{
\item{seu}{Seurat object}

\item{split.by}{The grouping variable used to split the input object. Default: NULL.}

\item{cluster.name}{Name of cluster field in Seurat object. Default: NULL.}
}
\value{
seurat object including a corrected counts matrix and contamination rates for each cell.
}
\description{
A wrapper of celda::decontX
}
